last function2

+++ src/main/java/com/example/demo/Services/Impl/EventServiceImpl.java
@@ -11,6 +11,10 @@ import com.example.demo.Repository.RegistrationRepository;
 import com.example.demo.Repository.UserRepository;
 import com.example.demo.Services.EventServices;
 import com.example.demo.User;
+import com.example.demo.exception.EventCannotBeDeletedException;
+import com.example.demo.exception.EventNonExistant;
+import com.example.demo.exception.RegistrationNonExistent;
+import com.example.demo.exception.UserNonExistent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -18,6 +22,7 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 
+import java.time.LocalDateTime;
 import java.util.*;
 import java.util.UUID;
 
@@ -29,7 +34,7 @@ public class EventServiceImpl implements EventServices {
     private final RegistrationRepository registrationRepository;
     private final UserRepository userRepository;
 
-    @Autowired // Ajoutez cette annotation pour l'injection de dépendances
+    @Autowired
     public EventServiceImpl(EventRepository repository, EventRepository eventRepository, RegistrationRepository registrationRepository, UserRepository userRepository) {
         this.repository = repository;
         this.eventRepository = eventRepository;
@@ -54,8 +59,7 @@ public class EventServiceImpl implements EventServices {
 
     @Override
     public Event updateEvent(UUID EventID, EventDTOUpdate EventDTO){
-        Event event = repository.findById(EventID).orElse(null);
-        assert event != null;
+        Event event = repository.findById(EventID).orElseThrow(()-> new EventNonExistant("Event not found"));
         event.setTitle(EventDTO.getTitle());
         event.setDescription(EventDTO.getDescription());
         event.setEventStatus(EventDTO.getEventStatus());
@@ -67,6 +71,14 @@ public class EventServiceImpl implements EventServices {
 
     @Override
     public void deleteEvent(UUID idEvent) {
+        Event event = repository.findById(idEvent).orElseThrow(()-> new EventNonExistant("Event not found with id" + idEvent.toString()));
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime date_end = event.getDate_end();
+
+        if(date_end.isBefore(now)){
+            throw new EventCannotBeDeletedException("Event cannot be deleted because it is over");
+        }
+
         repository.deleteById(idEvent);
     }
 
@@ -77,6 +89,7 @@ public class EventServiceImpl implements EventServices {
 
     @Override
     public Event getEventById(UUID idEvent) {
+        Event event = repository.findById(idEvent).orElseThrow(()-> new EventNonExistant("Event not found with id" + idEvent.toString()));
         return repository.findById(idEvent).orElse(null);
     }
 
@@ -89,11 +102,11 @@ public class EventServiceImpl implements EventServices {
 
     @Override
     public List<Event> getEventByMember(UUID memberId) {
-        User user = userRepository.findById(memberId).orElse(null);
-        if (user == null) {
-            return null;
-        }
+        User user = userRepository.findById(memberId).orElseThrow(()-> new UserNonExistent("User not found"));
         List<Registration> registrations = registrationRepository.findByUser(user);
+        if (registrations.isEmpty()) {
+            throw new RegistrationNonExistent("No registration for this member");
+        }
         List<Event> events = new ArrayList<>();
         for (Registration registration : registrations) {
             events.add(registration.getEvent());
@@ -108,15 +121,14 @@ public class EventServiceImpl implements EventServices {
 
     @Override
     public List<Event> getEventByNonRegisteredMember(UUID memberId) {
-        User user = userRepository.findById(memberId).orElse(null);
-        if (user == null) {
-            return null;
-        }
+        User user = userRepository.findById(memberId).orElseThrow(()-> new UserNonExistent("User not found"));;
         List<Registration> registrations = registrationRepository.findByUser(user);
+        if (registrations.isEmpty()) {
+            throw new RegistrationNonExistent("No registration for this member");
+        }
         List<Event> allEvents = eventRepository.findAll();
         List<Event> nonRegisteredEvents = new ArrayList<>();
 
-        // Vérifie pour chaque événement s'il n'existe pas d'enregistrement correspondant à l'utilisateur
         for (Event event : allEvents) {
             boolean isRegistered = false;
             for (Registration registration : registrations) {
@@ -134,6 +146,10 @@ public class EventServiceImpl implements EventServices {
 
         Set<Event> uniqueEvents = new LinkedHashSet<>(nonRegisteredEvents);
         List<Event> uniqueEventsList = new ArrayList<>(uniqueEvents);
+
+        if(uniqueEventsList.isEmpty()){
+            throw new EventNonExistant("No events found for this Id");
+        }
 
         return uniqueEventsList;
     }
