+++ src/main/java/com/example/demo/Services/Impl/RegistrationServiceImpl.java
@@ -8,6 +8,10 @@ import com.example.demo.Repository.RegistrationRepository;
 import com.example.demo.Repository.UserRepository;
 import com.example.demo.Services.RegistrationServices;
 import com.example.demo.User;
+import com.example.demo.exception.EventNonExistant;
+import com.example.demo.exception.RegistrationNonExistent;
+import com.example.demo.exception.UserAlreadyRegisteredException;
+import com.example.demo.exception.UserNonExistent;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
@@ -31,8 +35,16 @@ public class RegistrationServiceImpl implements RegistrationServices {
 
     @Override
     public Registration CreateRegistration(RegistrationDTOCreate registrationDTO) {
-        User user = UserRepository.findById(registrationDTO.getUserId()).orElse((null));
-        Event event = EventRepository.findById(registrationDTO.getEventId()).orElse((null));
+        User user = UserRepository.findById(registrationDTO.getUserId()).orElseThrow(()-> new UserNonExistent("User not found with id: " + registrationDTO.getUserId()));
+        Event event = EventRepository.findById(registrationDTO.getEventId()).orElseThrow(()-> new EventNonExistant("Event not found with id: " + registrationDTO.getEventId()));
+
+
+        List<Registration> allRegistration = RegistrationRepository.findAll();
+        for(Registration r : allRegistration){
+            if((r.getEvent().getIdEvent() == event.getIdEvent()) && (r.getUser().getId() == user.getId())) {
+                throw new UserAlreadyRegisteredException("User already registered");
+            }
+        }
         Registration registration = new Registration();
         registration.setEvent(event);
         registration.setUser(user);
@@ -41,6 +53,9 @@ public class RegistrationServiceImpl implements RegistrationServices {
 
     @Override
     public void deleteRegistration(UUID id_Registration) {
+        if (!RegistrationRepository.existsById(id_Registration)) {
+            throw new RegistrationNonExistent("Registration not found with id: " + id_Registration);
+        }
         RegistrationRepository.deleteById(id_Registration);
     }
 
@@ -51,6 +66,6 @@ public class RegistrationServiceImpl implements RegistrationServices {
 
     @Override
     public Registration getRegistrationById(UUID id_Registration) {
-        return RegistrationRepository.findById(id_Registration).orElse(null);
+        return RegistrationRepository.findById(id_Registration).orElseThrow(()-> new RegistrationNonExistent("Registration not found with id: " + id_Registration));
     }
 }
